add `JobCard` which is used in `JobPanel`
